django_flutter_auth_jwt/requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
PyJWT
pytz
sqlparse
tzdata
pyotp
google-api-python-client
django-jazzmin